Update `emitCmp` and optimize binary operations handling
Changed `emitCmp` to support 64-bit immediate values consistently across implementations. Refactored binary operation logic to simplify handling of immediate values and leverage unified pathways for improved maintainability.

+++ src/Backend/SCBE/Encoder/SCBE_X64.h
@@ -41,7 +41,7 @@ struct SCBE_X64 : SCBE_CPU
 
     void emitCmp(CPUReg reg0, CPUReg reg1, CPUBits numBits);
     void emitCmp(CPUReg memReg, uint32_t memOffset, CPUReg reg, CPUBits numBits);
-    void emitCmp(CPUReg memReg, uint32_t memOffset, uint32_t value, CPUBits numBits);
+    void emitCmp(CPUReg memReg, uint32_t memOffset, uint64_t value, CPUBits numBits);
     void emitCmp(CPUReg reg, uint64_t value, CPUBits numBits);
     void emitSet(CPUReg reg, CPUSet setType);
     void emitClear(CPUReg reg, CPUBits numBits);
